Add support for chaging databits length
With this option PHiLIP supports following combinations:

* 7E1, 7E2
* 7O1, 7O2
* 8N1, 8N2
* 8E1, 8E2
* 8O1, 8O2

Mask MSB for 7-bit mode as otherwise the characters can be
misinterpreted when MSB is "1".

As pointed in some forums [1] UART behaves differently in DMA mode. If
parity error occurs DMA won't put this character into a buffer, so only
the bytes with correct parity will get into the userspace.

[1] https://community.st.com/s/question/0D50X00009XkeWfSAJ/stm32f3-uart-dma-problem

Signed-off-by: Yegor Yefremov <yegorslists@googlemail.com>