fix(fw/pinout): Change SPI pins to pullup
The bug causes floating spi pins, nss and sck to trigger an interrupt.
This locks the PHiLIP up trying to handle an interrupt that cannot be handled.
The SPI TXE and BUSY flags are set and cannot be unset.
By changing all the inputs (NSS, MOSI, SCK) to pullup, floating pins no longer trigger interrupts.
It also appeared that the NSS pin for the nucleo was not initialized properly so that was corrected.