fix(fw/spi): Harden SPI communication
It turns out the MISO pin was creating issues on the SCK.
The transistion from PHiLIP are too fast creating ringing on the SCK corrupting data.
This error only occured with specific wiring setups, sometimes it was fine other times not.
By setting the  it slows down the MISO enough to prevent the problem.
A setup could reproduce the error when GPIO_SPEED_FREQ_HIGH and could not when GPIO_SPEED_FREQ_MEDIUM.
GPIO_SPEED_FREQ_MEDIUM states it is limited to 10 MHz vs GPIO_SPEED_FREQ_HIGH which is 50 MHz.
Since RIOT only support up to 10 MHz GPIO_SPEED_FREQ_MEDIUM is the best option.