feat(fw/tmr): Implementation of timer module
This allows for very precise and accurate capture of pin toggles.

This module requires very low jitter since the precision is 14ns. In interrupt mode events can occur as fast as 1us apart, however, time must be allowed for the buffer to be processed. With the DMA it appears that events can still be caught up to 200ns. However the buffer still must not overflow too quickly.

Some cleanup has also been done to the other timer msp stuff.