feat(qual/adc): Add adc qualfication test
This uses the Digilent Analog Discovery 2 to generate varying DC levels.
PHiLIPs DUT_ADC pin is connected to the W1 pin on the DAD2.
It tests various voltages in both slow and fast sampling modes.
It verifies the variance of reads after PHiLIP resets to test offset calibration.
It also verifies the accuracy.
Currently the offset tolarance is 25mV and the accuracy is 10mV.